refactor: move dimensional reduction logic into reddwarf-ts
Extracts types, config, and a pure `runReducer()` generator from the
app into the reddwarf-ts library package, enabling the polis pipeline
to run in JavaScript outside the browser (e.g. Node.js scripts).

The app worker and React hook become thin shells; no consuming files
needed import changes. `@saehrimnir/druidjs` added as a reddwarf-ts
runtime dependency.
