+++ packages/reddwarf-ts/src/index.ts
@@ -1,2 +1,3 @@
 export * from './stats.js';
 export * from './representative-statements.js';
+export * from './druid-reducer.js';
